Adjust to upcoming Julia 1.12 change in implicit world age increment
In 1.12, we'll be cleaning up and making consistent when world age increment happens. This package is currently relying on implicit world age increments that may go away in 1.12. See JuliaLang/julia#56509. This PR adds appropriate annotations that the world age increment is required after the include. This PR strictly retains existing behavior. A potentially better solution may be to use invokelatest and/or move more work out of the macro, which would also allow this code to work at non-toplevel (which may nor may not be desirable).